+++ eva/src/components/dashBoardInside/dashGuntt.vue
@@ -3,24 +3,27 @@
     class="dash-guntt"
     :class="idDashClass()"
   >
-    <div class="legend-block">
+    <div
+      v-if="!noMsg && guntt && guntt.getPhases.length > 0"
+      class="legend-block"
+    >
       <div
-        v-for="i in legends.length"
-        v-show="!noMsg"
-        :key="i"
+        v-for="(item, index) in guntt.getPhases"
+        :key="`${item}${index}`"
         class="legends-itself"
       >
         <div
           class="circle"
-          :style="{ backgroundColor: colors[i - 1] }"
+          :style="{ backgroundColor: colors[index] }"
         />
         <div class="text">
-          {{ legends[i - 1] }}
+          {{ item }}
         </div>
       </div>
     </div>
     <div
-      v-show="!noMsg"
+      v-if="!noMsg"
+      ref="gunttBlock"
       class="guntt-block"
     />
     <div
@@ -32,7 +35,7 @@
       }"
     />
     <div
-      v-show="noMsg"
+      v-if="noMsg"
       class="mt-4"
     >
       {{ msgText }}
@@ -41,7 +44,7 @@
 </template>
 
 <script>
-import * as d3 from 'd3';
+import GunttClass from '../../js/classes/GunttClass';
 
 export default {
   props: {
@@ -100,6 +103,7 @@ export default {
       ],
       legends: [],
       firstTime: true, // определяем первый ли раз зашли на страницу, ничего лучше не придумал
+      guntt: null,
     };
   },
   computed: {
@@ -124,17 +128,8 @@ export default {
           if (this.dataReport) {
             if (this.activeElemFrom === this.id) {
               this.prepareChart(this.dataRestFrom);
-            } else {
-              // .dash-guntt-${this.id}
-              const graphics = d3
-                .select(this.$el.querySelector('.guntt-block'))
-                .selectAll('svg')
-                .nodes(); // получаем область в которой будем рисовтаь график
-
-              if (graphics.length !== 0) {
-                // если график уже есть
-                graphics[0].remove(); // удаляем его
-              }
+            } else if (this.guntt) {
+              this.guntt.removeGuntt();
             }
           } else {
             this.prepareChart(this.dataRestFrom);
@@ -154,16 +149,8 @@ export default {
           if (this.dataReport) {
             if (this.activeElemFrom === this.id) {
               this.prepareChart(this.dataRestFrom);
-            } else {
-              const graphics = d3
-                .select(this.$el.querySelector('.guntt-block'))
-                .selectAll('svg')
-                .nodes(); // получаем область в которой будем рисовтаь график
-
-              if (graphics.length !== 0) {
-                // если график уже есть
-                graphics[0].remove(); // удаляем его
-              }
+            } else if (this.guntt) {
+              this.guntt.removeGuntt();
             }
           } else {
             this.prepareChart(this.dataRestFrom);
@@ -184,16 +171,8 @@ export default {
           if (this.dataReport) {
             if (this.activeElemFrom === this.id) {
               this.prepareChart(this.dataRestFrom);
-            } else {
-              const graphics = d3
-                .select(this.$el.querySelector('.guntt-block'))
-                .selectAll('svg')
-                .nodes(); // получаем область в которой будем рисовтаь график
-
-              if (graphics.length !== 0) {
-                // если график уже есть
-                graphics[0].remove(); // удаляем его
-              }
+            } else if (this.guntt) {
+              this.guntt.removeGuntt();
             }
           } else {
             this.prepareChart(this.dataRestFrom);
@@ -207,16 +186,8 @@ export default {
           if (this.dataReport) {
             if (this.activeElemFrom === this.id) {
               this.prepareChart(this.dataRestFrom);
-            } else {
-              const graphics = d3
-                .select(this.$el.querySelector('.guntt-block'))
-                .selectAll('svg')
-                .nodes(); // получаем область в которой будем рисовтаь график
-
-              if (graphics.length !== 0) {
-                // если график уже есть
-                graphics[0].remove(); // удаляем его
-              }
+            } else if (this.guntt) {
+              this.guntt.removeGuntt();
             }
           } else {
             this.prepareChart(this.dataRestFrom);
@@ -246,16 +217,8 @@ export default {
               if (this.activeElemFrom === this.id) {
                 this.noMsg = false;
                 this.prepareChart(this.dataRestFrom);
-              } else {
-                const graphics = d3
-                  .select(this.$el.querySelector('.guntt-block'))
-                  .selectAll('svg')
-                  .nodes(); // получаем область в которой будем рисовтаь график
-
-                if (graphics.length !== 0) {
-                  // если график уже есть
-                  graphics[0].remove(); // удаляем его
-                }
+              } else if (this.guntt) {
+                this.guntt.removeGuntt();
               }
             } else {
               this.noMsg = false;
@@ -303,391 +266,125 @@ export default {
       });
     },
     createChart(sizeChart, that, dataRest) {
-      let otstupBot = 30;
-      if (window.screen.width <= 1600) {
-        otstupBot = 10;
-      }
-
       const margin = {
-        top: 20, right: 20, bottom: otstupBot, left: 20,
+        top: 20,
+        right: 20,
+        bottom: window.screen.width <= 1600 ? 10 : 30,
+        left: 20,
+      };
+      const paddings = {
+        right: 80,
+        left: window.screen.width > 1920 ? 90 : 70,
+      };
+      const timeAxisStyles = {
+        height: window.screen.width > 1920 ? 60 : 50,
+        paddingTop: window.screen.width > 1920 ? 15 : 10,
+        paddingLeft: window.screen.width < 1400 ? -10 : 0,
       };
       const width = sizeChart.width - margin.left - margin.right;
       const height = sizeChart.height - margin.top - margin.bottom;
-      let otstupLeft = 70;
-      const otstupRight = 80;
-
-      if (window.screen.width > 1920) {
-        otstupLeft = 90;
-      }
+      const dateFormat = this.timeFormatFrom || '%Y-%m-%d %H:%M:%S';
 
       const data = [];
       dataRest.forEach((item) => {
         data.push({ ...{}, ...item });
       });
-      const graphics = d3
-        .select(this.$el.querySelector('.guntt-block'))
-        .selectAll('svg')
-        .nodes(); // получаем область в которой будем рисовтаь график
-
-      if (graphics.length !== 0) {
-        // если график уже есть
-        graphics[0].remove(); // удаляем его
+      if (this.guntt) {
+        this.guntt.removeGuntt();
       }
-
-      const svg = d3
-        .select(this.$el.querySelector('.guntt-block'))
-        .append('svg')
-        .attr('width', width)
-        .attr('height', height)
-        .attr('class', 'guntt-svg');
-
-      function checkZero(item) {
-        if (item < 10) {
-          // если там больше 10 символов
-          item = `0${item}`;
-        }
-        return item;
-      }
-
-      data.forEach((item, i) => {
-        let newDate = new Date(item.start_date * 1000);
-        data[i].start_date = `${newDate.getFullYear()}-${checkZero(
-          newDate.getMonth() + 1,
-        )}-${checkZero(newDate.getDate())} ${checkZero(
-          newDate.getHours(),
-        )}:${checkZero(newDate.getMinutes())}:${checkZero(
-          newDate.getSeconds(),
-        )}`;
-        newDate = new Date(item.end_date * 1000);
-        data[i].end_date = `${newDate.getFullYear()}-${checkZero(
-          newDate.getMonth() + 1,
-        )}-${checkZero(newDate.getDate())} ${checkZero(
-          newDate.getHours(),
-        )}:${checkZero(newDate.getMinutes())}:${checkZero(
-          newDate.getSeconds(),
-        )}`;
+      const guntt = new GunttClass({
+        elem: this.$refs.gunttBlock,
+        width,
+        height,
+        margin,
+        textColor: this.colorFrom.text || 'white',
+        colors: this.colors,
+        paddings,
+        timeAxisStyles,
+        data,
+        dateFormat,
+        tooltipElem: this.$refs.tooltip,
       });
+      this.guntt = Object.freeze(guntt);
 
-      let dateFormat = '%Y-%m-%d %H:%M:%S';
-      if (that.timeFormatFrom !== '') {
-        dateFormat = that.timeFormatFrom;
+      if (this.guntt.getPhases.length > 0) {
+        this.legends = this.guntt.getPhases;
       }
-
-      const x = d3
-        .scaleTime()
-        .domain([
-          d3.min(data, (d) => new Date(d.start_date)),
-          d3.max(data, (d) => new Date(d.end_date)),
-        ])
-        .range([otstupLeft, width - otstupRight]);
-
-      // рассчитываем набор фаз (состояний)
-      const phases = [];
-      const ids = [];
-      data.forEach((ph) => {
-        if (ph.phase) {
-          if (!phases.includes(ph.phase)) {
-            phases.push(ph.phase);
-          }
-        }
-        if (!ids.includes(ph.id)) {
-          ids.push(ph.id);
-        }
-      });
-      if (phases.length > 0) {
-        this.legends = phases;
-      }
-
-      let otstupBottom = 50;
-
-      if (window.screen.width > 1920) {
-        otstupBottom = 60;
-      }
-
-      const barHeight = Math.round((height - otstupBottom) / ids.length);
-
-      let otstupX = 0;
-
-      if (window.screen.width < 1400) {
-        otstupX = -10;
-      }
-
-      let deliter = 2;
-
-      if (width + margin.left + margin.right < 600) {
-        deliter = 3;
-      }
-      if (width + margin.left + margin.right < 450) {
-        deliter = 4;
-      }
-      if (width + margin.left + margin.right < 400) {
-        deliter = 5;
-      }
-      if (width + margin.left + margin.right < 350) {
-        deliter = 6;
-      }
-
-      function wrap(text) {
-        text.each(function () {
-          const localText = d3.select(this);
-          let row = [];
-          localText
-            .node()
-            .querySelectorAll('text')
-            .forEach((item) => {
-              row = item.innerHTML.split(' ').filter((rowItem) => rowItem !== '');
-              if (row[1]) {
-                row[0] = `<tspan x=0 y=10>${row[0]}</tspan>`;
-                row[1] = `<tspan x=0 y=30>${row[1]}</tspan>`;
-                item.innerHTML = row.join('');
-              }
-            });
-        });
-      }
-
-      // добавляем ось X
-      const xAxis = svg
-        .append('g')
-        .attr('transform', `translate(0,${height - otstupBottom})`)
-        .call(
-          d3
-            .axisBottom(x)
-            .tickFormat(d3.timeFormat(dateFormat))
-            .tickValues(
-              x.ticks().filter((item, i) => i % deliter === 0),
-            ),
-        )
-        .call(wrap);
-
-      let otstupY = 10;
-
-      if (window.screen.width > 1920) {
-        otstupY = 15;
-      }
-
-      xAxis.selectAll('line').attr('y2', `-${height}`).attr('opacity', '0.3');
-      xAxis
-        .selectAll('text')
-        .attr('class', 'text-x')
-        .attr('transform', `translate(${otstupX},${otstupY})`)
-        .style('color', this.colorFrom.text)
-        .style('text-anchor', 'center');
-
-      // горизонатьные значения (id)
-      svg
-        .append('g')
-        .selectAll('rect')
-        .data(ids)
-        .enter()
-        .append('rect')
-        .attr('x', 0)
-        .attr('y', (d, i) => i * barHeight)
-        .attr('width', () => width)
-        .attr('height', barHeight)
-        .attr('stroke', 'none')
-        .attr('opacity', 0.2);
-
-      // сами строки данных относителньо времени
-      const bars = svg.append('g').selectAll('rect').data(data).enter();
-      const lines = bars
-        .append('rect')
-        .attr('rx', 3)
-        .attr('ry', 3)
-        .attr('x', (d, index) => x(Date.parse(d.start_date)) - (index % 2 === 0 ? 1 : 0))
-        .attr('y', (d) => {
-          let j = -1;
-          ids.forEach((item, i) => {
-            if (item === d.id) {
-              j = i;
-            }
-          });
-          return j * barHeight - (j === 0 ? 1 : 0);
-        })
-        .attr('width', (d) => x(Date.parse(d.end_date)) - x(Date.parse(d.start_date)))
-        .attr('height', barHeight)
-        .attr('stroke', 'none')
-        .style('cursor', 'pointer')
-        .attr('fill', (d) => {
-          let j = 0;
-          phases.forEach((item, i) => {
-            if (item === d.phase) {
-              j = i;
-            }
-          });
-
-          return this.colors.filter(Boolean)[j];
-        });
 
       // Tooltip
-
       const tooltipBlock = this.$refs.tooltip;
       const tooltipMargin = this.$attrs['is-full-screen'] ? 200 : 30;
 
       function transformDescription(text) {
         let rows = text.split('\\n');
         rows = rows.map((item) => `<p class="row-toolrip">${item}</p>`);
-        // rows = '<div class = "tooltip-guntt">' + rows.join('') + '</div>';
         return rows.join('');
       }
 
-      function moveTooltip(offsetX) {
-        const localY = d3.event.offsetY;
-        const localX = d3.event.offsetX + offsetX;
-        tooltipBlock.style.top = `${localY}px`;
-        tooltipBlock.style.left = `${localX}px`;
-      }
+      this.guntt.setEventsOnLines([
+        {
+          name: 'mouseover',
+          callback: (event) => {
+            let tooltip = '';
 
-      lines
-        .on('mouseover', (event) => {
-          let tooltip = '';
-
-          if (data[0].description) {
-            tooltip = transformDescription(event.description);
-          } else {
-            Object.keys(event).forEach((key) => {
-              tooltip += `<p class="row-toolrip"><span>${key}</span>: ${event[key]}</p>`;
-            });
-          }
-          moveTooltip(tooltipMargin);
-          tooltipBlock.innerHTML = tooltip;
-          tooltipBlock.style.opacity = '0.9';
-          tooltipBlock.style.visibility = 'visible';
-        })
-        .on('mousemove', () => {
-          moveTooltip(tooltipMargin);
-        })
-        .on('mouseout', () => {
-          tooltipBlock.style.opacity = '0';
-          tooltipBlock.style.visibility = 'hidden';
-        })
-        .on('click', (d) => that.setClick(d));
-
-      const texts = bars
-        .append('text')
-        .text((d) => {
-          if (d.phase) {
-            if (
-              x(Date.parse(d.end_date)) - x(Date.parse(d.start_date))
-              > d.phase.length * 8
-            ) {
-              return d.phase;
+            if (data[0].description) {
+              tooltip = transformDescription(event.description);
+            } else {
+              Object.keys(event).forEach((key) => {
+                tooltip += `<p class="row-toolrip"><span>${key}</span>: ${event[key]}</p>`;
+              });
             }
-          }
-          return '';
-        })
-        .attr('x', (d) => (
-          x(Date.parse(d.start_date))
-            + (x(Date.parse(d.end_date)) - x(Date.parse(d.start_date))) / 2
-        ))
-        .attr('y', (d) => {
-          let j = -1;
-          ids.forEach((item, i) => {
-            if (item === d.id) {
-              j = i;
-            }
-          });
-          return j * barHeight + barHeight / 2 + 3;
-        })
-        .attr('class', 'text-bar')
-        .attr('font-size', 11)
-        .attr('text-anchor', 'middle')
-        .style('cursor', 'pointer')
-        .attr('fill', '#fff');
+            this.guntt.moveTooltip(tooltipMargin);
+            tooltipBlock.innerHTML = tooltip;
+            tooltipBlock.style.opacity = '0.9';
+            tooltipBlock.style.visibility = 'visible';
+          },
+        },
+        {
+          name: 'mousemove',
+          callback: () => {
+            this.guntt.moveTooltip(tooltipMargin);
+          },
+        },
+        {
+          name: 'mouseout',
+          callback: () => {
+            tooltipBlock.style.opacity = '0';
+            tooltipBlock.style.visibility = 'hidden';
+          },
+        },
+        {
+          name: 'click',
+          callback: (d) => that.setClick(d),
+        },
+      ]);
 
-      texts
-        .on('mouseover', () => {
-          tooltipBlock.style.opacity = '0.9';
-          tooltipBlock.style.visibility = 'visible';
-        })
-        .on('mousemove', () => {
-          moveTooltip(tooltipMargin);
-        })
-        .on('mouseout', () => {
-          tooltipBlock.style.opacity = '0';
-          tooltipBlock.style.visibility = 'hidden';
-        })
-        .on('click', (d) => that.setClick(d));
-
-      // подписи слева
-
-      let currentPos = 0;
-
-      const idsCaption = svg.append('g').selectAll('text').data(ids).enter();
-
-      function checkCaption(name) {
-        if (name.length > 6) {
-          // если там больше 10 символов
-          name = `${name.substring(0, 6)}...`; // обрезаем и добовляем троеточие
-        }
-        return name;
-      }
-
-      idsCaption
-        .append('text')
-        .text((d) => checkCaption(d))
-        .attr('x', 10)
-        .attr('y', (d, i) => {
-          if (i === 0) {
-            currentPos = barHeight / 2 + 2;
-          } else {
-            currentPos += barHeight + 2;
-          }
-          return currentPos;
-        })
-        .attr('class', 'ids-caption')
-        .attr('font-size', 11)
-        .attr('text-anchor', 'start')
-        .style('opacity', '0.8')
-        .attr('fill', this.colorFrom.text || 'white');
-
-      currentPos = 0;
-
-      idsCaption
-        .append('line')
-        .attr('x1', 0)
-        .attr('x2', width - otstupRight)
-        .attr('y1', (d, i) => {
-          let curPos;
-          if (i !== ids.length - 1) {
-            curPos = currentPos + barHeight;
-            currentPos += barHeight;
-          } else {
-            curPos = currentPos + barHeight;
-            currentPos = 0;
-          }
-          return curPos;
-        })
-        .attr('y2', () => {
-          const curPos = currentPos + barHeight;
-          currentPos += barHeight;
-          return curPos;
-        })
-        .style('stroke', this.colorFrom.text)
-        .attr('opacity', '0.3');
-
-      // линия отделяющяя подписи (ids) от блока данных
-
-      svg
-        .append('g')
-        .append('line')
-        .attr('x1', otstupLeft)
-        .attr('x2', otstupLeft)
-        .attr('y1', 0)
-        .attr('y2', height - 50)
-        .style('stroke', 'black')
-        .attr('opacity', '0.3');
-
-      // линия отделяющяя блоки данных от легенды
-
-      svg
-        .append('g')
-        .append('line')
-        .attr('x1', width - otstupRight)
-        .attr('x2', width - otstupRight)
-        .attr('y1', 0)
-        .attr('y2', height - 50)
-        .style('stroke', 'black')
-        .attr('opacity', '0.3');
+      this.guntt.setEventsOnTexts([
+        {
+          name: 'mouseover',
+          callback: () => {
+            tooltipBlock.style.opacity = '0.9';
+            tooltipBlock.style.visibility = 'visible';
+          },
+        },
+        {
+          name: 'mousemove',
+          callback: () => {
+            this.guntt.moveTooltip(tooltipMargin);
+          },
+        },
+        {
+          name: 'mouseout',
+          callback: () => {
+            tooltipBlock.style.opacity = '0';
+            tooltipBlock.style.visibility = 'hidden';
+          },
+        },
+        {
+          name: 'click',
+          callback: (d) => that.setClick(d),
+        },
+      ]);
     },
 
     setClick(item) {