+++ CHANGELOG.md
@@ -9,15 +9,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unrelise]
 
 ### Feature
-- Added hotkey for query window in data research v2
-- created map class
+- Created map class
 
 ### Fixed
+- fixed maps mode
+
+## [1.88.0]
+
+### Feature
+- Added hotkey for query window in data research v2
+- Created guntt class
+
+### Fixed
+- fix export link
 - fixed data exploration 2.0 gantt: the date format setting for the Gantt chart does not apply
 - fixed change osmserver in map settings
 - fixed default osmserver and default search in dashMapUserSettings
 - fixed maps tooltip
-- fixed maps mode
+- fixed the display of histograms with negative values on a multiline dash
+- the width of the histogram is now only positive
 
 ## [1.87.0]